+++ controllers/clonar_calendario_academico.go
@@ -1,12 +1,8 @@
 package controllers
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-
 	"github.com/astaxie/beego"
-	"github.com/udistrital/utils_oas/request"
+	"github.com/udistrital/sga_mid_calendario_academico/services"
 	"github.com/udistrital/utils_oas/requestresponse"
 
 	"github.com/udistrital/utils_oas/errorhandler"
@@ -32,162 +28,18 @@ func (c *ClonarCalendarioController) URLMapping() {
 func (c *ClonarCalendarioController) PostCalendario() {
 	defer errorhandler.HandlePanic(&c.Controller)
 
-	var calendario map[string]interface{}
-	var calendarioParam []map[string]interface{}
-	var tipoEvento []map[string]interface{}
-	var calendarioEvento []map[string]interface{}
-	var calendarioEventoTipoPublico []map[string]interface{}
-	var tipoPublico map[string]interface{}
-	var resultadoPost map[string]interface{}
-	var resultadoPostResponsable map[string]interface{}
-	var errCalendarioParam = errors.New("")
-	var errorGetAll bool
-	var message string
+	data := c.Ctx.Input.RequestBody
 
-	var dataPost map[string]interface{}
-	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &dataPost); err == nil {
+	resultado, err := services.PostCalendario(data)
 
-		idCalendario := fmt.Sprintf("%.f", dataPost["Id"].(float64))
-		idPeriodo := fmt.Sprintf("%.f", dataPost["PeriodoIdClone"].(float64))
-		idNivel := fmt.Sprintf("%.f", dataPost["NivelClone"].(float64))
-		c.Data["json"] = idCalendario
-
-		errCalendario := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario/"+idCalendario, &calendario)
-		if errCalendario == nil {
-			if calendario != nil {
-
-				// if dataPost["NivelClone"].(float64) == calendario["Nivel"].(float64) {
-				// 	errCalendarioParam = request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario?query=Activo:true,PeriodoId:"+idPeriodo+",Nivel:"+idNivel+"&sortby=Id&order=desc&offset=1&limit=0", &calendarioParam)
-				// } else {
-				// 	errCalendarioParam = request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario?query=Activo:true,PeriodoId:"+idPeriodo+",Nivel:"+idNivel+"&sortby=Id&order=desc", &calendarioParam)
-				// }
-
-				errCalendarioParam = request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario?query=Activo:true,PeriodoId:"+idPeriodo+",Nivel:"+idNivel+"&sortby=Id&order=desc", &calendarioParam)
-
-				if errCalendarioParam == nil {
-					if calendarioParam != nil && calendarioParam[0]["Id"] != nil {
-
-						idCalendarioParam := fmt.Sprintf("%.f", calendarioParam[0]["Id"].(float64))
-
-						// persistir tipo_evento si el calendario que se esta clonando los tiene
-						errTipoEvento := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"tipo_evento?query=CalendarioID__Id:"+idCalendarioParam, &tipoEvento)
-						if errTipoEvento == nil {
-							if tipoEvento != nil && tipoEvento[0]["Id"] != nil {
-								for _, tEvento := range tipoEvento {
-
-									idOld := fmt.Sprintf("%.f", tEvento["Id"].(float64))
-									tEvento["Id"] = 0
-									tEvento["CalendarioID"] = calendario
-
-									errTipoEventoPost := request.SendJson("http://"+beego.AppConfig.String("EventoService")+"/tipo_evento", "POST", &resultadoPost, tEvento)
-									if errTipoEventoPost == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-										if resultadoPost["Status"] != 400 {
-											tEvento["Id"] = resultadoPost["Id"]
-
-											// presistir calendario_evento si el tipo_evento que se esta clonando esta asociado en el campo tipo_evento_id del calendario_evento
-											errCalendarioEvento := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario_evento?query=TipoEventoId__Id:"+idOld, &calendarioEvento)
-											if errCalendarioEvento == nil {
-												if calendarioEvento != nil && calendarioEvento[0]["Id"] != nil {
-													idCalendarioEventoOld := fmt.Sprintf("%.f", calendarioEvento[0]["Id"].(float64))
-													for _, cEvento := range calendarioEvento {
-
-														cEvento["Id"] = 0
-														cEvento["TipoEventoId"] = tEvento
-														cEvento["FechaInicio"] = "2020-01-01T00:00:00-05:00"
-														cEvento["FechaFin"] = "2020-01-01T00:00:00-05:00"
-
-														errCalendarioEventoPost := request.SendJson("http://"+beego.AppConfig.String("EventoService")+"/calendario_evento", "POST", &resultadoPost, cEvento)
-														if errCalendarioEventoPost == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-															if resultadoPost["Status"] != 400 {
-
-																//validar si existe relcion de responsables, tabla rompimiento calendario_evento_tipo_publico y tipo_publico
-																errCalendarioEventoTipoPublico := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario_evento_tipo_publico?query=CalendarioEventoId__Id:"+idCalendarioEventoOld, &calendarioEventoTipoPublico)
-																if errCalendarioEventoTipoPublico == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-																	if resultadoPost["Status"] == nil {
-																		for _, cEventoTipoPublico := range calendarioEventoTipoPublico {
-																			tipoPublicoOld := fmt.Sprintf("%.f", cEventoTipoPublico["TipoPublicoId"].(map[string]interface{})["Id"].(float64))
-																			errTipoPublico := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"tipo_publico/"+tipoPublicoOld, &tipoPublico)
-																			if errTipoPublico == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-																				if resultadoPost["Status"] == nil {
-
-																					cEventoTipoPublico["Id"] = 0
-																					cEventoTipoPublico["CalendarioEventoId"] = resultadoPost
-																					cEventoTipoPublico["TipoPublicoId"] = tipoPublico
-
-																					errCalendarioEventoPost := request.SendJson("http://"+beego.AppConfig.String("EventoService")+"/calendario_evento_tipo_publico", "POST", &resultadoPostResponsable, cEventoTipoPublico)
-																					if errCalendarioEventoPost == nil && fmt.Sprintf("%v", resultadoPostResponsable["System"]) != "map[]" && resultadoPostResponsable["Id"] != nil {
-																						if resultadoPost["Status"] != 400 {
-																							// fmt.Println("calendario_evento nuevo: ", resultadoPostResponsable["Id"])
-																						}
-																					}
-
-																				}
-																			} else {
-																				errorGetAll = true
-																				message += errTipoPublico.Error()
-																			}
-																		}
-																	}
-																} else {
-																	errorGetAll = true
-																	message += errCalendarioEventoTipoPublico.Error()
-																}
-
-															}
-														} else {
-															errorGetAll = true
-															message += errCalendarioEventoPost.Error()
-														}
-
-													}
-												}
-											} else {
-												errorGetAll = true
-												message += errCalendarioEvento.Error()
-											}
-										}
-									} else {
-										errorGetAll = true
-										message += errTipoEventoPost.Error()
-									}
-
-								}
-
-							} else {
-								errorGetAll = true
-								message += "No data found tipoEvento[0]"
-							}
-						} else {
-							errorGetAll = true
-							message += errTipoEvento.Error()
-						}
-					} else {
-						errorGetAll = true
-						message += "No data found calendarioParam[0]"
-					}
-				} else {
-					errorGetAll = true
-					message += errCalendarioParam.Error()
-				}
-			} else {
-				errorGetAll = true
-				message += "No data found"
-			}
-		} else {
-			errorGetAll = true
-			message += errCalendarioParam.Error()
-		}
-
-	}
-	if !errorGetAll {
+	if err == nil {
 		c.Ctx.Output.SetStatus(200)
-		response := requestresponse.APIResponseDTO(true, 200, calendario)
-		c.Data["json"] = response
+		c.Data["json"] = requestresponse.APIResponseDTO(true, 200, resultado)
 	} else {
-		c.Ctx.Output.SetStatus(400)
-		response := requestresponse.APIResponseDTO(false, 400, nil, message)
-		c.Data["json"] = response
+		c.Ctx.Output.SetStatus(404)
+		c.Data["json"] = requestresponse.APIResponseDTO(true, 404, nil, err.Error())
 	}
+
 	c.ServeJSON()
 
 }
@@ -202,141 +54,18 @@ func (c *ClonarCalendarioController) PostCalendario() {
 func (c *ClonarCalendarioController) PostCalendarioPadre() {
 	defer errorhandler.HandlePanic(&c.Controller)
 
-	var calendario map[string]interface{}
-	var calendarioParam []map[string]interface{}
-	var tipoEvento []map[string]interface{}
-	var calendarioEvento []map[string]interface{}
-	var resultadoPost map[string]interface{}
-	var resultado map[string]interface{}
-	var errCalendarioParam = errors.New("")
-	var errorGetAll bool
-	var message string
-	var statusCode int = 200
+	data := c.Ctx.Input.RequestBody
 
-	var dataPost map[string]interface{}
-	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &dataPost); err == nil {
-		idCalendario := fmt.Sprintf("%.f", dataPost["Id"].(float64))
-		idCalendarioPadre := fmt.Sprintf("%.f", dataPost["IdPadre"].(map[string]interface{})["Id"])
-		errCalendario := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario/"+idCalendario, &calendario)
-		if errCalendario == nil {
-			if calendario != nil {
-				if dataPost["Nivel"].(float64) == calendario["Nivel"].(float64) {
-					errCalendarioParam = request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario?query=Id:"+idCalendarioPadre, &calendarioParam)
-				} else {
-					errCalendarioParam = request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario?query=Id:"+idCalendarioPadre, &calendarioParam)
-				}
+	resultado, err := services.PostCalendarioPadre(data)
 
-				if errCalendarioParam == nil {
-					if calendarioParam != nil && calendarioParam[0]["Id"] != nil {
-						idCalendarioParam := fmt.Sprintf("%.f", calendarioParam[0]["Id"].(float64))
-
-						// persistir tipo_evento si el calendario que se esta clonando los tiene
-						errTipoEvento := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"tipo_evento?query=CalendarioID__Id:"+idCalendarioParam, &tipoEvento)
-						if errTipoEvento == nil {
-							if tipoEvento != nil && tipoEvento[0]["Id"] != nil {
-								for _, tEvento := range tipoEvento {
-									idOld := fmt.Sprintf("%.f", tEvento["Id"].(float64))
-									tEvento["Id"] = 0
-									tEvento["CalendarioID"] = calendario
-
-									errTipoEventoPost := request.SendJson("http://"+beego.AppConfig.String("EventoService")+"/tipo_evento", "POST", &resultadoPost, tEvento)
-									if errTipoEventoPost == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-										if resultadoPost["Status"] != 400 {
-											tEvento["Id"] = resultadoPost["Id"]
-
-											// presistir calendario_evento si el tipo_evento que se esta clonando esta asociado en el campo tipo_evento_id del calendario_evento
-											errCalendarioEvento := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario_evento?query=TipoEventoId__Id:"+idOld, &calendarioEvento)
-											if errCalendarioEvento == nil {
-												if calendarioEvento != nil && calendarioEvento[0]["Id"] != nil {
-													for _, cEvento := range calendarioEvento {
-														cEvento["Id"] = 0
-														cEvento["TipoEventoId"] = tEvento
-														cEvento["FechaInicio"] = "2000-01-01T00:00:00-05:00"
-														cEvento["FechaFin"] = "2000-01-01T00:00:00-05:00"
-
-														errCalendarioEventoPost := request.SendJson("http://"+beego.AppConfig.String("EventoService")+"/calendario_evento", "POST", &resultadoPost, cEvento)
-														if errCalendarioEventoPost == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-															if resultadoPost["Status"] != 400 {
-																fmt.Println("calendario_evento nuevo: ", resultadoPost["Id"])
-															} else {
-																errorGetAll = true
-																statusCode = 400
-																message += errCalendarioEventoPost.Error()
-															}
-														} else {
-															errorGetAll = true
-															statusCode = 404
-															message += "No data found"
-														}
-													}
-												} else {
-													errorGetAll = true
-													statusCode = 400
-													message += errCalendarioEvento.Error()
-												}
-											} else {
-												errorGetAll = true
-												statusCode = 404
-												message += "No data found"
-											}
-										} else {
-											errorGetAll = true
-											statusCode = 400
-											message += errTipoEventoPost.Error()
-										}
-									} else {
-										errorGetAll = true
-										statusCode = 404
-										message += "No data found"
-									}
-
-								}
-								resultado = map[string]interface{}{
-									"Id": idCalendario,
-								}
-							} else {
-								errorGetAll = true
-								statusCode = 404
-								message += "No data found"
-							}
-						} else {
-							errorGetAll = true
-							statusCode = 400
-							message = errTipoEvento.Error()
-						}
-					} else {
-						errorGetAll = true
-						statusCode = 404
-						message += "No data found"
-					}
-				} else {
-					errorGetAll = true
-					statusCode = 400
-					message += errCalendarioParam.Error()
-				}
-			} else {
-				errorGetAll = true
-				statusCode = 404
-				message += "No data found"
-			}
-		} else {
-			errorGetAll = true
-			statusCode = 400
-			message += errCalendario.Error()
-		}
-	} else {
-		errorGetAll = true
-		statusCode = 400
-		message += err.Error()
-	}
-
-	if !errorGetAll {
-		c.Ctx.Output.SetStatus(statusCode)
+	if err == nil {
+		c.Ctx.Output.SetStatus(200)
 		c.Data["json"] = requestresponse.APIResponseDTO(true, 200, resultado)
 	} else {
-		c.Ctx.Output.SetStatus(statusCode)
-		c.Data["json"] = requestresponse.APIResponseDTO(false, statusCode, nil, message)
+		c.Ctx.Output.SetStatus(404)
+		c.Data["json"] = requestresponse.APIResponseDTO(true, 404, nil, err.Error())
 	}
+
 	c.ServeJSON()
 }
 
@@ -350,156 +79,16 @@ func (c *ClonarCalendarioController) PostCalendarioPadre() {
 func (c *ClonarCalendarioController) PostCalendarioExtension() {
 	defer errorhandler.HandlePanic(&c.Controller)
 
-	var CalendarioExtension map[string]interface{}
-	var CalendarioPadre map[string]interface{}
-	var CalendarioExtCrear map[string]interface{}
+	data := c.Ctx.Input.RequestBody
 
-	var tipoEvento []map[string]interface{}
-	var calendarioEvento []map[string]interface{}
-	var calendarioEventoTipoPublico []map[string]interface{}
-	var tipoPublico map[string]interface{}
-	var resultadoPost map[string]interface{}
-	var resultadoPostResponsable map[string]interface{}
+	resultado, err := services.PostCalendarioExtension(data)
 
-	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &CalendarioExtension); err == nil {
-		IdPadre := fmt.Sprintf("%.f", CalendarioExtension["CalendarioPadre"])
-
-		errCalendarioPadre := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario/"+IdPadre, &CalendarioPadre)
-		if errCalendarioPadre == nil {
-			if _, ok := CalendarioPadre["Id"]; ok {
-				CalendarioPadre["AplicaExtension"] = true
-				CalendarioPadre["DependenciaParticularId"] = CalendarioExtension["Dependencias"]
-				CalendarioPadre["DocumentoExtensionId"] = CalendarioExtension["DocumentoExtensionId"]
-				CalendarioPadre["CalendarioPadreId"] = map[string]interface{}{"Id": CalendarioExtension["CalendarioPadre"]}
-				delete(CalendarioPadre, "Id")
-
-				errCalendarioExtCrear := request.SendJson("http://"+beego.AppConfig.String("EventoService")+"calendario", "POST", &CalendarioExtCrear, CalendarioPadre)
-				if errCalendarioExtCrear == nil {
-
-					// persistir tipo_evento si el calendario que se esta clonando los tiene
-					errTipoEvento := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"tipo_evento?query=CalendarioID__Id:"+IdPadre+",Activo:true&limit=0", &tipoEvento)
-					if errTipoEvento == nil {
-						if tipoEvento != nil && tipoEvento[0]["Id"] != nil {
-							for _, tEvento := range tipoEvento {
-
-								idOld := fmt.Sprintf("%.f", tEvento["Id"].(float64))
-								tEvento["Id"] = 0
-								tEvento["CalendarioID"] = CalendarioExtCrear
-
-								errTipoEventoPost := request.SendJson("http://"+beego.AppConfig.String("EventoService")+"/tipo_evento", "POST", &resultadoPost, tEvento)
-								if errTipoEventoPost == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-									if resultadoPost["Status"] != 400 {
-										tEvento["Id"] = resultadoPost["Id"]
-
-										// presistir calendario_evento si el tipo_evento que se esta clonando esta asociado en el campo tipo_evento_id del calendario_evento
-										errCalendarioEvento := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario_evento?query=TipoEventoId__Id:"+idOld+",Activo:true&limit=0", &calendarioEvento)
-										if errCalendarioEvento == nil {
-											if calendarioEvento != nil && calendarioEvento[0]["Id"] != nil {
-												idCalendarioEventoOld := fmt.Sprintf("%v", calendarioEvento[0]["Id"])
-												for _, cEvento := range calendarioEvento {
-
-													cEvento["Id"] = 0
-													cEvento["TipoEventoId"] = tEvento
-
-													errCalendarioEventoPost := request.SendJson("http://"+beego.AppConfig.String("EventoService")+"/calendario_evento", "POST", &resultadoPost, cEvento)
-													if errCalendarioEventoPost == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-														if resultadoPost["Status"] != 400 {
-
-															//validar si existe relcion de responsables, tabla rompimiento calendario_evento_tipo_publico y tipo_publico
-															errCalendarioEventoTipoPublico := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"calendario_evento_tipo_publico?query=CalendarioEventoId__Id:"+idCalendarioEventoOld+",Activo:true&limit=0", &calendarioEventoTipoPublico)
-															if errCalendarioEventoTipoPublico == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-																if resultadoPost["Status"] == nil {
-																	for _, cEventoTipoPublico := range calendarioEventoTipoPublico {
-																		tipoPublicoOld := fmt.Sprintf("%.f", cEventoTipoPublico["TipoPublicoId"].(map[string]interface{})["Id"].(float64))
-																		errTipoPublico := request.GetJson("http://"+beego.AppConfig.String("EventoService")+"tipo_publico/"+tipoPublicoOld, &tipoPublico)
-																		if errTipoPublico == nil && fmt.Sprintf("%v", resultadoPost["System"]) != "map[]" && resultadoPost["Id"] != nil {
-																			if resultadoPost["Status"] == nil {
-
-																				cEventoTipoPublico["Id"] = 0
-																				cEventoTipoPublico["CalendarioEventoId"] = resultadoPost
-																				cEventoTipoPublico["TipoPublicoId"] = tipoPublico
-
-																				errCalendarioEventoPost := request.SendJson("http://"+beego.AppConfig.String("EventoService")+"/calendario_evento_tipo_publico", "POST", &resultadoPostResponsable, cEventoTipoPublico)
-																				if errCalendarioEventoPost == nil && fmt.Sprintf("%v", resultadoPostResponsable["System"]) != "map[]" && resultadoPostResponsable["Id"] != nil {
-																					if resultadoPost["Status"] != 400 {
-																						// fmt.Println("calendario_evento nuevo: ", resultadoPostResponsable["Id"])
-																					}
-																				}
-
-																			}
-																		} else {
-																			//errorGetAll = true
-																			fmt.Println("error get tipo_publico", errTipoPublico)
-																			c.Ctx.Output.SetStatus(404)
-																			c.Data["json"] = requestresponse.APIResponseDTO(false, 404, nil, errTipoPublico.Error())
-																		}
-																	}
-
-																	c.Ctx.Output.SetStatus(200)
-																	c.Data["json"] = requestresponse.APIResponseDTO(true, 200, nil)
-
-																}
-															} else {
-																//errorGetAll = true
-																fmt.Println("error get calend_evento_tipo_pub", errCalendarioEventoTipoPublico)
-																c.Ctx.Output.SetStatus(404)
-																c.Data["json"] = requestresponse.APIResponseDTO(false, 404, nil, errCalendarioEventoTipoPublico.Error())
-															}
-														}
-													} else {
-														//errorGetAll = true
-														fmt.Println("error post calend event", errCalendarioEventoPost)
-														c.Ctx.Output.SetStatus(400)
-														c.Data["json"] = requestresponse.APIResponseDTO(false, 400, nil, errCalendarioEventoPost.Error())
-													}
-
-												}
-											}
-										} else {
-											//errorGetAll = true
-											fmt.Println("error get calend_evento", errCalendarioEvento)
-											c.Ctx.Output.SetStatus(404)
-											c.Data["json"] = requestresponse.APIResponseDTO(false, 404, nil, errCalendarioEvento.Error())
-										}
-									}
-								} else {
-									//errorGetAll = true
-									fmt.Println("error put evento", errTipoEventoPost)
-									c.Ctx.Output.SetStatus(404)
-									c.Data["json"] = requestresponse.APIResponseDTO(false, 404, nil, errTipoEventoPost.Error())
-								}
-
-							}
-
-						} else {
-							c.Ctx.Output.SetStatus(200)
-							c.Data["json"] = requestresponse.APIResponseDTO(true, 200, nil, "ok pero no eventos")
-						}
-					} else {
-						fmt.Println("error get eventos", errTipoEvento)
-						c.Ctx.Output.SetStatus(404)
-						c.Data["json"] = requestresponse.APIResponseDTO(false, 404, nil, errTipoEvento.Error())
-					}
-
-				} else {
-					fmt.Println("error put calend padre extension", errCalendarioExtCrear)
-					c.Ctx.Output.SetStatus(400)
-					c.Data["json"] = requestresponse.APIResponseDTO(false, 400, nil, errCalendarioExtCrear.Error())
-				}
-			} else {
-				fmt.Println("error get calend padre")
-				c.Ctx.Output.SetStatus(404)
-				c.Data["json"] = requestresponse.APIResponseDTO(false, 404, nil)
-			}
-		} else {
-			fmt.Println("error get calend padre", errCalendarioPadre)
-			c.Ctx.Output.SetStatus(404)
-			c.Data["json"] = requestresponse.APIResponseDTO(false, 404, nil, errCalendarioPadre.Error())
-		}
+	if err == nil {
+		c.Ctx.Output.SetStatus(200)
+		c.Data["json"] = resultado
 	} else {
-		fmt.Println("error body")
 		c.Ctx.Output.SetStatus(404)
-		c.Data["json"] = requestresponse.APIResponseDTO(false, 404, nil)
+		c.Data["json"] = requestresponse.APIResponseDTO(true, 404, nil, err.Error())
 	}
 
 	c.ServeJSON()
